Alter nullable constraint on created_at column
oslo_db takes care of populating 'created_at' column with valid data
at the time of inserting record so it's guaranteed that 'created_at'
column will be never null. So there is no need of 'nullable=false'
for it.

This patch alters the nullable constraint from 'False' to 'True' for
existing hosts and failover_segments table by adding a new migration
script.

Closes-Bug: #1623526

Change-Id: I9ce46b43796b790c318e244d722ac0825e36c1cc