Add admin role for masakari user
Masakari API's are allowed to be execute by admin only. So when external
service like masakari-instancemonitor calls masakari-api using masakari
user it does not have admin role associated with it's tenant. So it
returns HTTPForbidden 403 response to masakari-instancemonitor.

Added admin role to service tenant which has masakari user associated
with.

Change-Id: I505ea56bc022aad086f1090ff25b59ec85b70c41