Fix bad request error for valid marker in hosts
Even if we pass a valid host id as marker to the curl command
to return list of hosts, it returns a Bad Request error.
This is due to the use of 'FailoverSegment' instead of 'Host'
while generating the query.

This patch fixes the issue by replacing 'FailoverSegment' with
'Host'.

Updated test cases for failover segments and notifications along
with hosts for testing scenario with valid limit and marker.

Change-Id: I3c64d32f93a2bc09f1754b0fbb93d53c375899a2
Closes-bug: #1654154