Remove deps under pep8 section to enable hacking checks
Hacking checks are not working due to the presence of
"deps = hacking" under pep8 section. This installs the
latest version of hacking library which is 0.13.0 and
it has some issues due to which custom hacking checks
are not working.

After removing it from pep8 section, the hacking
version to be used is referred from test-requirements.txt
and it installs hacking library version 0.10.2 which does
not have this issue.

Change-Id: I6ef802838c05f93b15a3a5a0359b6ea3579fa476