InstanceFailure: Ignore notifications for certain events
For instance failure recovery Masakari doesn't check for the event from
payload received in notification, as a result it executes workflow for
all the events received.

When a instance is stopped (normal operation), masakari-instancemonitor
sends notifications related to following two events one after another to
masakari-api service.
    A) event: LIFECYCLE detail: STOPPED_SHUTDOWN
    B) event: LIFECYCLE detail: SHUTDOWN_FINISHED

When masakari-engine processes the first notification, it checks for vmstate
of the instance, if it’s 'stopped', it won't stop the vm. But if it's not
stopped it will attempt to stop it. Then again, libvirt will send the above
two events to masakari-instancemonitor and masakari-instancemonitor will send
related notifications to masakari-api and this way it will go into the recursion.

To solve this issue we can only process the below mentioned events at masakari
side which actually requires actions to be taken:
    A) 'event': 'LIFECYCLE', 'vir_domain_event': 'STOPPED_FAILED'
    B) 'event': 'IO_ERROR', 'vir_domain_event': 'IO_ERROR_REPORT'

So removed handling of all other events. For all other events masakari will
simply log the message at INFO level and mark the notification status as
'ignored'.

Closes-Bug: #1662994
Change-Id: If1113a599894cc7bab32b9656b6c42de5e5d7a54