Stop logging unnecessary warning on context create
Every time a message is dispatched over rpc, it serializes the context
using to_dict() method which includes "is_admin_project" key in the
serialized context which when deserialized on the receiving end calls
"from_dict" class method to build RequestContext object based on the
dictionary. At the time of creation of this RequestContext object, a
warning message is logged "Arguments dropped when creating context".

Passed "is_admin_project" to __init__() of base RequestContext class so
that the warning message will not come. Also this log message should be
logged in debug level instead of warning as it's meant for developers and
not operators. Developers can choose to see what arguments are getting
dropped so that if required later it can be passed to base RequestContext.
Kept the default value of "is_admin_project" as True [1].

[1] https://github.com/openstack/oslo.context/blob/master/oslo_context/context.py#L196

TrivialFix

Change-Id: Ie301908fd037d08998d85c46f9da567a9448fdd2