Change nova_catalog_admin_info to default to publicURL
A recent change in devstack [1] reduced the number of endpoints
configured by default. This breaks masakari resulting into:
"EndpointNotFound: adminURL endpoint for compute service not found"

This patch changes the default for the nova_catalog_admin_info to be
the normal 'publicURL'. With 'publicURL' masakari-engine is able to
call nova's admin restricted apis.

[1] Ided7a65c81b3a0b56f0184847fc82e17c29a771e
Closes-Bug: #1669280
Change-Id: I90bc214249f8a542f2d40fca544af204e0665821