Exclude on maintenance reserved_hosts for host_failure recovery
masakari-engine considers reserved_hosts which are under maintenance
for host failure recovery when recovery_method is reserved_host,
auto_priority or rh_priority and passes it to workflow for processing.

If the reserved_host is under maintenance then it should not be used
for evacuating the instances for host_failure recovery.

This patch excludes reserved_hosts which are on maintenance for
host_failure recovery by filtering host_list with
'on_maintenance'=False.

Closes-Bug: #1697902
Change-Id: Ie6beff110d03b67d6f5d8694603816c0db017660