Remove unused methods
Removed following methods as these methods are not called from anywhere
in the entire code:

 * is_user_context
 * require_admin_context
 * require_context
 * authorize_project_context
 * authorize_user_context

'is_user_context' method [1] was getting used to check whether the
requested context is of normal user or not. Masakari doesn't allow
normal user to call any API so the code was unreachable. Moreover
masakari don't have any sqlalchemy model having 'project_id' as it's
attribute. So it is safe to remove this code.

[1] https://github.com/openstack/masakari/blob/master/masakari/db/sqlalchemy/api.py#L131-L136

Change-Id: I60f68cbdd80dbce9d3ea95441cf1926ebec1cc8c