Modify unit testcases according to oslo.context
oslo.context>=2.19.1 outputs 'project' key in context's to_dict
function.
The value of 'project' key is same as 'project_id' key.

So this patch modifies unit testcases according to the above changs.

And, this patch adds oslo.context to requirements.txt.
oslo.context had been installed related to oslo.versioned object
until now.
But, it is better to describe it explicitly because Masakari uses
the library.

Change-Id: Ibebed8a4e5bfd49f192b1cf876b7538f9c057f97