Ship files like other openstack projects
For packagers, it's useful to have consistency.
This patch ensures the api-paste is shipped with masakari as
data files, like other projects do: [1], [2], [3].

In this patch, I also added the policy.json file, because
the policy is not yet in code.

Configuration files are not to be shipped with this manner,
as they are generated.

[1]: https://github.com/openstack/cinder/blob/master/setup.cfg#L28
[2]: https://github.com/openstack/glance/blob/master/setup.cfg#L20
[3]: https://github.com/openstack/nova/blob/master/setup.cfg#L26

Change-Id: I67aaca1da57909746766dc495b2009fb60feffd0