Fix Context test cases
oslo.context has added a new key 'system_scope" and hence some of the
unit tests are failing as it asserts on the entire contents of the
dict. Modified all such unit tests and now it only asserts on
masakari-specific keys in the context.

This patch removes test_convert_from_dict_to_dict_version_2_4_x unit
test as Masakari is using oslo.context>=2.19.2.

Change-Id: I3252ef742f5dd4cb446b1dd6cd887e42f2a9d729
Signed-off-by: Sampath Priyankara <sam47priya@gmail.com>
Co-Authored-By: Neha Alhat <neha.alhat@nttdata.com>