Recover resized instance(power_state=SHUTDOWN) with vm_state=stopped
The instance which is vm_state=resized and power_state=SHUTDOWN
should not be recovered with vm_state=error. It should be recovered
with vm_state=stopped. The instances such as paused or suspended
(which are other than vm_state=active,stopped,error) are also the same.

This patch modified that masakari-engine will recover these instances
with vm_state=stopped.

Change-Id: Ibdea705f69d139834c9b2294ae11bc9344c259bd
Closes-Bug: #1782518