Allow adding host only if it exists in nova
Presently, you can add any hostname to the failover segment. It
doesn't check whether this host exists or not in nova. This patch
fixes this issue by checking whether the hostname exists in nova or
not before adding it to the failover segment. If it doesn't exists,
it will raise 400 error.

APIImpact
BadRequest(400) is returned if host doesn't exists in nova instead
of 200 during host create.

Closes-Bug: #1800073

Change-Id: I0ccc9f3a61e3f17f2cb7c1ad1888947c6fe724c8