Allow updating host name only if it exists in nova
Presently, you can update host name with any non-existing host name.
It doesn't check whether this host exists or not in nova. This patch
fixes this issue by checking whether the host name exists in nova or
not before updating it. If it doesn't exists, it will raise 400 error.

APIImpact
BadRequest(400) is returned if host doesn't exists in nova instead of
200 during host update.

Closes-Bug: #1814656

Change-Id: Ibd113f2328deae0f1114544436631bdc434eff92