Check if host belongs to a segment
If you query a host passing an existing failover segment but not
the one that is assigned to the host, it still returns the host
successfully. In this case, it should fail with 404 error.

This patch checks if the host belongs to the segment that is
present in the URI. If not, it will return 404 error.

Change-Id: I16256cc2a01696a1d54cb9326aed17b723b87727
Closes-Bug: #1854323