handle duplicate notifications without checking status
If a notification is received and processed before getting a duplicate
notification within the time interval defined in config option
'duplicate_notification_detection_interval', it will be processed again.

In this case, it should be ignored, because whether notification is duplicate
is no matter with its status.

Duplicate notification is no matter with its status.

Change-Id: I56c5f72ebb3cefe2d36b3688c46e0b27d9b6ba7a
Close-Bug: #1858548