Search in nova services instead of hypervisors
Nova services and hypervisor naming can differ, as they retireve node
names in different way.
In the meanwhile we operate with nova.services while enabling/disabling
nodes duringh the incident. So we're supposed to have in database record
matching to what we have in service list, but not in hypervisor list.

Closes-Bug: #1839715
Change-Id: I9c591d33f17a8d5950bdb1fc2d686e2301fc6d95