Fix functional.test_hosts.test_list
A recent OpenStack SDK change [1] has caused more attributes to be
present in listings.
The logic in the failing test required 1:1 mapping between two
values so it started failing.
I have learnt from the OpenStack SDK (and mordred in particular)
that the contract is to not cease to return attributes but
openstacksdk is free to add as many extra ones as they see fit.
Hence, this changes makes the test more reasonable by following
that contract.

[1] https://review.opendev.org/c/openstack/openstacksdk/+/776205

Change-Id: Ife8b645c41cf681fa30564e75f7757f96c0bfe11