[CI] Fix the functional job
Make it compatible with the DevStack switch to OVN.

It cleans up the job definition to reuse defaults as
much as possible to avoid similar issues in the future.
The Swift (as previously) and Cinder (since now) are
disabled as they are irrelevant to Masakari.

This patch also turns off bindep (not useful in
functional tests) and limits tox installation to
the controller node (where it runs from).

The DevStack orchestration has been moved to the ``run``
phase to avoid wasting time on (``pre-run`` phase)
retries when Masakari fails to install via DevStack.

Change-Id: I70a10eb29e1fbef2786c2e7229aa2c615f6e5a5e