[DevStack] Do not use the sample process list
... for the process monitor deployment.

That sample list manipulation was ugly and definitions there
did not really match the reality causing random hiccups in CI and
local DevStack deployments (mostly due to constant restarting of
instancemonitor).
Moreover, processmonitor is deprecated in Xena. [1]
The existing unit tests and the noop deployment should be enough
to keep it on the surface.

[1] https://review.opendev.org/c/openstack/masakari-monitors/+/806431

Change-Id: I1787206b1ca84b64686bfaa1d00f33dbfc2e3adb