[CI] Install dependencies for docs target
When tox 'docs' target is called, first it installs the dependencies
(listed in 'deps') in 'installdeps' phase, then it installs masakari
(with its requirements) in 'develop-inst' phase. In the latter case
'deps' is not used so that the constraints defined in 'deps' are not
used. This could lead to failures on stable branches when new packages
are released that break the build. To avoid this, the simplest solution
is to pre-install requirements, i.e. add requirements.txt to 'docs' tox
target.

Conflicts:
  tox.ini

NOTE(elod.illes): conflict is due to branch specific upper constraints
file link.

Change-Id: Ic5e214be240023f0e4f02cae3939e1d199ecb061
(cherry picked from commit aa13e4536eadbce8de35109f62a379e7e375bd02)