Fix tox4 issues
New tox version 4 has changed interpretation of tox.ini

The passenv parsing requires the list to be comma-separated.
Since the names are case-insensitive, duplicates are removed.

skipsdist is now a synonym of no_package and avoids the
installation of the package in all cases.
However, some environments require it (for setup metadata).

Fix allowlist_externals where appropriate.

Change-Id: Icb067e85e4f4874a52710d7053a2048e337305c8