add pyproject.toml to support pip 23.1
pip 23.1 removed the "setup.py install" fallback for projects that do
not have pyproject.toml and now uses a pyproject.toml which is vendored
in pip. To address that, this change adds the minimal pyproject.toml to
enable pbr to be properly used to build editable wheels.

See [1] and [2] for more details on the removal of the fallback support.

[1] https://pip.pypa.io/en/stable/news/#v23-1
[2] https://github.com/pypa/pip/issues/8368

Change-Id: I24a1b24ff249b0d2e9a1b350a40d146a7fc358ba
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>