Move test base into tests directory
The masakari.test module is used only by unit tests. Move it to
the test code directory so that users can more easily determine
that the file is used only in testing.

Also make sure that the test codes are omitted in coverage report,
because we do not care coverage of test code.

Change-Id: I2ffc75bd2efacf4a18f546700af05abb7368fdd3